wedge shape
to try to reuse the gametrayz cover on the base, I've changed the outer
walls to the sloped wedge sides of the original insert.  Added fillets
everywhere but because the wedge sides get too close in the corners I
had to reduce the fillet there to 0.6mm.

Checking fit with next partial/test print.